requirements.txt
Pillow
images2gif
envoy
six
